Chore: Move metadata from layout to page
